+++ kopy-compiler/test-data/box/no-nest-copy-update-no-it.kt
@@ -15,7 +15,7 @@ fun box(): String {
 
     val foo22 = foo02 copy2 {
         number.update2 {
-            42.also { setKopyableReference2(getKopyableReference2().copy(number = it)) }
+            42.also { _atomic2.lazySet(_atomic2.value.copy(number = it)) }
         }
     }
 
@@ -38,12 +38,6 @@ public interface Kopyable2<T> {
     public val _atomic2: AtomicRef<T>
 
     public fun _initKopyable2(): Kopyable2<T>
-
-    public fun getKopyableReference2(): T = _atomic2.value
-
-    public fun setKopyableReference2(value: T) {
-        _atomic2.lazySet(value)
-    }
 
     public infix fun copy2(copy2: T.() -> Unit): T {
         val kopyable: Kopyable2<T> = _initKopyable2()